update for Bro 2.4
Added parsing to remove undesirable charterers from Sguil alert subject and updated Bro intel.log logic to handle the new file field.  Still having an issue on startup and restart of Bro where the agent stops reading the intel.log and notice.log files till it is restarted.  Using "tail -F" so I'm unsure of what the problem is.  Also tried pointing the agent to the linked Bro log directory and the permanent Bro log directory but that did not solve the problem either.